feature: enable interaction with panel
This had to be done with kind of a hacky way. Basically click handler
has to be hooked on window because the cursor canvas is stretched
all over the viewport and the clicks won't propagate to panel UI.

That means I'm looking up buttons via their coordinates :) But I think
it is cheap operation because it will only happen on click.

Next step is to do the same for keyboard and gamepad. Those devices are
the reason why the code is complex.